Change variable tag in For Loop to an int variable

+++ Step5_ControlFlowStatements/ForLoop/src/ForLoop.java
@@ -10,7 +10,7 @@ public class ForLoop {
             System.out.println("Loop " + i + " Hello!");
         }
 
-        for (double i = 2; i <= 5; i++) {
+        for (int i = 2; i <= 5; i++) {
             System.out.println("10,000 at " + i + "% interest = " + calculateInterest(10000.0, i));
         }
     }
